test(127052): Correção de fixture
